refactor(routes/Menu): :recycle: Rename ContextMenuToggle to ContextMenuTrigger
